[schema] add new supervision contact status type, scheduled_date (Recidiviz/recidiviz-data#44082)
## Description of the change

this PR implements
[this](https://www.notion.so/recidiviz/Enum-Change-Request-Add-SCHEDULED-to-StateSupervisionContactStatus-2097889f4d19807f9531cd433cbb3895)
scr, which adds a `SCHEDULED` status to the
`[Normalized]StateSupervisionContact`. additionally, adds
`scheduled_date` as well to *hopefully* avoid accidental usage of this
entity downstream. we also enforce the following three invariants for
contacts with a `SCHEDULED` status:
- it has a non-null `scheduled_date`
 - it has a null `contact_date`
 - both `verified_employment` and `resulted_in_arrest` are null

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#44081

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 46743b3e4e607a1c1b46f55bc43744a4537c6146