[deploy] Add looker staging deploy script (Recidiviz/recidiviz-data#44572)
## Description of the change

This script is intended to be run from the recidiviz-data repo root, and
it will clone the looker repo into a temporary
directory, verify that the looker repo files are up-to-date with the
recidiviz-data commit being deployed or will sync
them if not by creating a new branch in the looker repo with the updated
files. Once the looker repo files are up-to-date,
it will create a new version tag in the looker repo, and point the
looker project to the new tag.

takes the same args as `base_deploy_to_staging.sh`

i don't want to block the normal deploy on the looker repo being out of
sync, so my idea is to have this run at the end of the deploy scripts.

i'm waiting on permissions to set up looker deploy hooks so the actual
looker deploy part will come later

as it stands i think adding this step to the deploy would be pretty
annoying for the on call because its annoying as a dev to also have to
open a `looker` PR + manually run content validation after you merge
generated changes into `recidiviz-data` so i would imagine that the repo
would be out of sync pretty frequently since the generated files so far
are being updated multiple times per week.

however, the world that i am wanting to move towards is
1. add a script that syncs the repos and opens a looker PR as part of
the CI suite for recidiviz-data PRs and comments the looker PR link onto
the recidiviz-data PR.
2. add lookml validation to a looker repo CI suite
3. add a post-merge CI job that fails if the repos are out of sync,
similar to what we do for unsafe BQ schema updates

then the desired workflow for `generated/` changes would be 1. merge
looker PR once recidiviz-data changes are approved 2. merge
recidiviz-data PR. we could also add some CI check on recidiviz-data PRs
that fails until the looker repo PR is merged if we wanted.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#36190

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fe4196d0ae4884bb93a9dca2d1fe4fa5928cf538