[US_MO] Ingest 1270 (STG-related) assessment (Recidiviz/recidiviz-data#44651)
## Description of the change

Creates a new assessment ingest view in MO that will be used to ingest
all assessments that aren't ORAS/classification assessments, which each
have their own views. Currently, this is only needed for 1270
assessments: this is a recently-added assessment type that designates
STG affiliation, which is part of the criteria for work release in MO.

Includes some minor tweaks to various raw data configs that needed join
logic or had incorrect flags for PII fields to support fixture
generation for the ingest view test.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d37b309256e6d15205895b8220b3bdd0f03d231b