[US_TN] Ingest Diversion Sentences (Recidiviz/recidiviz-data#44453)
## Description of the change

This PR ingests charges, sentences, and statuses for diversion sentences
and introduces a couple of TODOs for more clarity and accuracy.

1. [Recidiviz/recidiviz-data#44429](https://github.com/Recidiviz/recidiviz-data/issues/44429) to
update sentences that have judgement order/sentence information for a
diversion sentence _of the same ID_! My guess is that this is for
revocations, but I am not certain yet. To account for this now, we
append "DIVERSION" to IDs here so they do not merge and break (basically
what we do in V1). There is a test case for this!
2. [Recidiviz/recidiviz-data#44447](https://github.com/Recidiviz/recidiviz-data/issues/44447) to
further understand and document what diversion type is in our data.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Supports
[Recidiviz/recidiviz-data#44414](https://github.com/Recidiviz/recidiviz-data/issues/44414)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a2352ca18baf390a286fdc37ba988be9b73d98c7