[US_TX][Insights] Update vitals metrics query for MONTH period metrics (Recidiviz/recidiviz-data#44818)
## Description of the change

To launch the Operations module of the Supervisor Homepage in Texas, we
will use the new due-date-based timely contacts metric, which will have
values for the 1st of the current and previous months, rather than
yesterday and 30 days ago - see Recidiviz/recidiviz-data#44473 for the expected schema.

This PR updates `_get_vitals_metrics_from_officer_pseudonymized_ids` to
account for this, namely:
- Use 1 month ago, rather than 30 days ago, as the cutoff date for
selecting the metric from the "current period" when the `period` is
"MONTH". This makes a difference on certain edge case dates, like near
the end of February or at the end of 31-day months.
- Also send to the frontend the date on which the current and previous
metric values were calculated. AFAIK, we'll have two categories of
calculations in Texas Operations module - contact compliance will be
month-based and risk assessment compliance will be calculated on a
rolling 30-day basis - so I think having the option to display these in
some way will be helpful.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-dashboards/issues/8764

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e7e3b91add4bff196eac12a2951f0f389b812764