[US_ME][US_TN] Rename external_id to payment_account_external_id in fees/invoices views (Recidiviz/recidiviz-data#45013)
## Description of the change

There are a number of views related to payments / invoices which have an
`external_id` column. The name `external_id` is ambiguous (is it a
person external id? something else?). From my read of these views, this
`external_id` is associated with a specific "account" that the payments
are credited to and is a separate key we need to group on / partition by
in addition to `person_id`. As part of the work for Recidiviz/recidiviz-data#44757, this PR
renames all the columns in these fees / invoices views from
`external_id` to `payment_account_external_id`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#44757

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e9692720f1f903f59cab467ae9f1d51a5d656dd8