[monitoring] factor in task retries in airflow monitoring model (Recidiviz/recidiviz-data#44718)
## Description of the change

this PR accounts for task retries in our airflow monitoring model; this
is predicated upon the fact that airflow uses `TaskInstanceHistory` to
store all "non-current" (i.e. previous) task instance attempts.

i had to change a little bit of the logic in how we determine last/next
success because i couldn't get the way we were doing it to work with
sorting on two keys in a way that didnt make things way more complicated
-- any suggestions here would be MUCH appreciated if you have an opinion
about it!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25994

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Nick Tallant <150271010+recidinick@users.noreply.github.com>
GitOrigin-RevId: 24a986ff2fbc0c9f07d9afaadaf65a698a62eca0