[Admin Panel] Add Action Strategies default copy to Insights config form (Recidiviz/recidiviz-data#45030)
## Description of the change

I was having trouble updating the Texas insights config through the
admin panel. The current config has an empty dictionary for Action
Strategy Copy, so the `AddConfigForm` showed no form items for Action
Strategies, so upon submitting the form the resulting config didn't
contain the field `actionStrategyCopy`, so we got an error - we expect
this field to be defined.

This PR adds the default copy as a default, so that if there is no
config or the current config has an empty dictionary, the new config
form can still be submitted. From the insights db schema for
configurations, it seems like this copy is meant to be the default value
anyway.

Tested locally, I was able to add a config to a state with none defined.

I'm not sure how the config was added in the first place without having
this field set, so probably this could have been fixed for the TX case
by using that method that I'm not aware of? But it seems possibly
confusing for any new Supervisor Homepage states in the future.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 251f94785db3b9800aec9e77391d55e3172413e9