[Insights][Usage] Add latest login date to supervision officers query  (Recidiviz/recidiviz-data#44603)
## Description of the change

Preview:
[[LINK](https://console.cloud.google.com/bigquery?ws=!1m5!1m4!4m3!1srecidiviz-staging!2sken_latest_login_outliers_views!3ssupervision_officers_materialized)]

This update introduces a new Common Table Expression (CTE) to the
supervision officers query, which retrieves the latest login date for
each officer. The CTE normalizes email addresses and aggregates the
maximum login date, allowing for a left join with the supervision
officers data. The latest login date is then included in the final
selection of columns.

Also add latest login date to SupervisionOfficerEntity and update
related queries and tests

## Type of change

- Type: Feature | non-breaking change that adds functionality

## Related issues

## Checklists

### Development

- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

### Code review

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 592b7d1d0c015a117d9de4d9052d2d49b1ceb4d6