[sftp] add channel timeout to prevent indefinite hang (Recidiviz/recidiviz-data#45181)
## Description of the change

@emilyemilyemilyemilyemilyemily saw an sftp task runtime alert in
missouri and noticed that there has been a few files that were just
hanging indefinitely until they hit the 8 hour execution timeout. this
PR adds an additional backstop against this by setting a socket
(paramiko's channel class) read/write timeout on the packet level of 3
minutes. paramiko very kindly notes that
<img width="598" alt="Screenshot 2025-07-09 at 12 32 58 PM"
src="https://github.com/user-attachments/assets/924ab550-34dd-4c95-8593-1028fe1fbd9f"
/>
so it internally limits any given request made to the channel to that
size. the timeout parameter we are passing to the channel is used
whenever reading that amount of data, and _does not_ apply to the
overall read time of the file. here are the docs from paramiko's channel
class about the timeout

```
        Set a timeout on blocking read/write operations.  The ``timeout``
        argument can be a nonnegative float expressing seconds, or ``None``.
        If a float is given, subsequent channel read/write operations will
        raise a timeout exception if the timeout period value has elapsed
        before the operation has completed.  Setting a timeout of ``None``
        disables timeouts on socket operations.
```

however, i'm not sure of the likelihood that this will fix the issues,
as the paramiko's buffered pipe class (where the timeout is passed to)
says something a little different about this timeout
```
        The optional ``timeout`` argument can be a nonnegative float expressing
        seconds, or ``None`` for no timeout.  If a float is given, a
        `.PipeTimeout` will be raised if the timeout period value has elapsed
        before any data arrives.
```

so it seems that this timeout will only prevent issues where we are
hanging w/ no data having been received (not sure what percentage of the
time that is the case vs part of the way through). in either case, hope
it'll be helpful! and we have 8 hour execution timeout to fallback on
which i think is a sensible value for now since sometimes files can take
forever. if we see more runtime alerts after this, let's consider moving
the execution timeout to the state-specific sftp delegates

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1e2ef2b3a24c8933f111fca8e136e54519fa291c