[Normalization] Autogenerate Snapshot IDs and create function to split them on a given status and time (Recidiviz/recidiviz-data#44684)
## Description of the change

This PR does two things:
1. Creates a helper function to split status snapshots on a given time
and status. This will be used on imposed pending serving status creation
when the parent sentence's terminating status falls in the middle of a
child serving status. (The test case is this scenario) Note we don't
actually add this in to the pipeline yet, as we need a little more
thought in the implementation.
2. Autogenerates the primary key for snapshots, which we now have to do
since we are potentially creating net-new statuses when splitting them
on serving starts.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#44573

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: bdca6f12792ac38e6f17fa15708b502dcd1ec704