[US_MI][Workflows] Add in clients who will be eligible in 30 days to upcoming tab for all supervision opps (Recidiviz/recidiviz-data#44548)
## Description of the change

Here we add almost eligible clients to early discharge, classification
review, and telephone reporting opportunities in MI.

**Eligibility changes**
- Classification review: There are more almost eligible clients for
classification review than eligible clients, so the number of clients in
the client record more than doubled.
- Early discharge: The client record now holds an extra 14% of clients
surfaced for ED that are almost eligible. Roughly 250 of these 1178
clients are almost eligible for parole ED and the rest are almost
eligible for probation ED.
- Telephone reporting: The client record now holds an extra 13% of
clients surfaced for TR that are almost eligible.
<img width="915" alt="image"
src="https://github.com/user-attachments/assets/b800411e-b6e0-49d8-b243-d50923df03d7"
/>

**Admin panel and FE changes required**
For CR:
- add eligible and almost eligible copy for both of the TES criteria.
(Leave the eligible copy for usMiClassificationReviewPastDueDate for the
moment.)
- Once that copy has been added we can just get rid of the frontend
criteria merging, since it's brittle and not too helpful. - We'll want
the frontend change to get to prod before the almost eligible clients
start showing up, but that shouldn't be a problem.

For ED and TR:
- I'll just update the admin panel copy accordingly!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-dashboards/issues/8417

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 37dde0686e92ff7083d39b3721faf66da62e618d